finished still to deploy

+++ controllers/categoryCtrl.js
@@ -1,4 +1,5 @@
 const Category = require("../models/categoryModel")
+const Products = require("../models/productModel")
 
 
 const categoryCtrl = {
@@ -6,6 +7,8 @@ const categoryCtrl = {
     getCategories: async(req,res)=>{
 
          try{
+
+           
 
             const categories = await Category.find()
             res.json(categories)
@@ -42,6 +45,8 @@ const categoryCtrl = {
 
     deleteCategory: async (req,res)=>{
 
+        const products = await Products.findOne({category:req.params.id})
+        if(products) return res.status(400).json({msg:"Please delete products which belong to this category."})
         try{
 
         await Category.findByIdAndDelete(req.params.id)
